fix: handle case when email includes partial provider domain name A bug was reported by a user when trying to sign in with his professional email address.
The user was using a professional email address, such as name@analytics.com, but he could not sign in as his email was rejected.

This email address was considered as a provider email address (and not a professional one) as it includes the letters cs.com; cs.com being listed as a provider domain.